Adapted test utilities to mocking mechanism
